backend/requirements.txt
Django==3.2.19
django-cors-headers==4.0.0
django-background-tasks==1.2.5
web3==6.4.0
requests==2.31.0
